# packages/locus-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Locus loopback API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/entities": {
      "get": {
        "tags": [
          "core"
        ],
        "operationId": "entity_ids",
        "parameters": [
          {
            "name": "X-Locus-Run",
            "in": "header",
            "description": "Expected backend run from private readiness; context, not authorization. Never silently replace it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Complete packed RFC UUIDv7 identities, 16 bytes each; unspecified sequence order. DB work completes before transfer. Empty success is zero bytes.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              },
              "Content-Length": {
                "schema": {
                  "type": "string"
                },
                "description": "Exact decimal byte count, required even for empty success"
              },
              "X-Content-Type-Options": {
                "schema": {
                  "type": "string"
                },
                "description": "nosniff"
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authorization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Wrong run or conflicting request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Operation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Admission closed or retained launch rejection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "core"
        ],
        "operationId": "create_entity",
        "parameters": [
          {
            "name": "X-Locus-Run",
            "in": "header",
            "description": "Expected backend run from private readiness; context, not authorization. Never silently replace it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestIdentity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationOutcome"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authorization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Wrong run or conflicting request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Operation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Admission closed or retained launch rejection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "task"
        ],
        "summary": "SSE `snapshot` events contain complete TaskSnapshot JSON, not a JSON response.",
        "operationId": "events",
        "parameters": [
          {
            "name": "X-Locus-Run",
            "in": "header",
            "description": "Expected backend run from private readiness; context, not authorization. Never silently replace it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream. Each snapshot event has TaskSnapshot JSON in data, and its revision in id. Subscribe with the authorization/run headers; reconnect establishes a fresh replacement snapshot.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authorization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Wrong run or conflicting request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Operation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Admission closed or retained launch rejection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/requests/{request_id}": {
      "get": {
        "tags": [
          "server"
        ],
        "operationId": "submission",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Locus-Run",
            "in": "header",
            "description": "Expected backend run from private readiness; context, not authorization. Never silently replace it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Submission"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authorization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Wrong run or conflicting request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Operation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Admission closed or retained launch rejection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/media-runtime": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "settings_runtime",
        "parameters": [
          {
            "name": "X-Locus-Run",
            "in": "header",
            "description": "Expected backend run from private readiness; context, not authorization. Never silently replace it.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRuntimeObservation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authorization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Wrong run or conflicting request ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Operation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Admission closed or retained launch rejection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "diagnostic": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DomainDiagnostic"
              }
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "unauthorized",
          "foreign_origin",
          "wrong_run",
          "invalid_request",
          "request_conflict",
          "admission_closed",
          "restricted",
          "launch_rejected",
          "unknown_request",
          "unknown_task",
          "missing_file",
          "missing_bytes",
          "access_denied",
          "preview_unavailable",
          "operation_failed",
          "not_found",
          "method_not_allowed"
        ]
      },
...
      "MediaRuntimeObservation": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "runtime",
              "status"
            ],
            "properties": {
              "runtime": {
                "$ref": "#/components/schemas/MediaSettingsRuntime"
              },
              "status": {
                "type": "string",
                "enum": [
                  "active"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "status"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "unavailable"
                ]
              }
            }
          }
        ]
      },
      "MediaSettingsRuntime": {
        "type": "object",
        "description": "Captured is the saved startup value; changes to saved values do not reconfigure\nthis run. Effective paths also include explicit process environment overrides.",
        "required": [
          "captured",
          "ffprobe",
          "ffmpeg"
        ],
        "properties": {
          "captured": {
            "$ref": "#/components/schemas/SavedSettings"
          },
          "ffmpeg": {
            "$ref": "#/components/schemas/EffectiveToolPath"
          },
          "ffprobe": {
            "$ref": "#/components/schemas/EffectiveToolPath"
          }
        }
      },
...
      "Receipt": {
        "type": "object",
        "required": [
          "run_id",
          "request_id",
          "task_id"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          },
          "run_id": {
            "type": "string"
          },
          "task_id": {
            "type": "string"
          }
        }
      },
...
      "RequestIdentity": {
        "type": "object",
        "required": [
          "request_id"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Submission": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "status"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "admission_pending"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "receipt",
              "status"
            ],
            "properties": {
              "receipt": {
                "$ref": "#/components/schemas/Receipt"
              },
              "status": {
                "type": "string",
                "enum": [
                  "accepted"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "error",
              "status"
            ],
            "properties": {
              "error": {
                "$ref": "#/components/schemas/ApiError"
              },
              "status": {
                "type": "string",
                "enum": [
                  "rejected"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "status"
            ],
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "direct_pending"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "outcome",
              "status"
            ],
            "properties": {
              "outcome": {
                "$ref": "#/components/schemas/MutationOutcome"
              },
              "status": {
                "type": "string",
                "enum": [
                  "direct_complete"
                ]
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}